web_app/requirements.txt
Flask
requests
yfinance
fastmcp
google-generativeai
PyJWT
